requirement.txt
torch>=1.9.0
numpy>=1.18.0
tqdm>=4.0.0
matplotlib
pykan
scikit-learn>=0.24.0
PyYAML
pandas
fdfdpy
cupy-cuda116